Simply code by hiding BigDecimal construction behind overloaded constructor.

+++ src/main/java/co/uk/mytutor/repository/InMemoryBookRepository.java
@@ -3,7 +3,6 @@ package co.uk.mytutor.repository;
 import co.uk.mytutor.model.Book;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -13,11 +12,11 @@ public class InMemoryBookRepository implements BookRepository {
     private final ConcurrentHashMap<String, Book> books = new ConcurrentHashMap<>();
 
     public InMemoryBookRepository() {
-        Book a = new Book("A", 10, BigDecimal.valueOf(25));
-        Book b = new Book("B", 10, BigDecimal.valueOf(20));
-        Book c = new Book("C", 10, BigDecimal.valueOf(23));
-        Book d = new Book("D", 10, BigDecimal.valueOf(30));
-        Book e = new Book("E", 10, BigDecimal.valueOf(27));
+        Book a = new Book("A", 10, 25);
+        Book b = new Book("B", 10, 20);
+        Book c = new Book("C", 10, 23);
+        Book d = new Book("D", 10, 30);
+        Book e = new Book("E", 10, 27);
 
         books.put(a.getName(), a);
         books.put(b.getName(), b);
